+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## TBD
+## 1.10
 
 ### Features
 
@@ -12,7 +12,7 @@
 * Backtesting stop orders were being canceled if the stop price was not hit.
 * Backtesting order updates were sometimes being processed after bar events. Order update events should get processed first.
 * Backtesting fills had the datetime of the event and now they have the datetime of the bar where the fill occurred.
-* Backtesting `OrderInfo.fill_price` was being calculated using the base and quote amounts filled and that could lead to invalid values.
+* Backtesting `OrderInfo.fill_price` was being calculated using the base and quote amounts filled, and that was sometimes leading to invalid values.
 
 ### Misc